requirements.txt
pandas==1.0.3
torchvision==1.5.0
tabulate==0.8.7